Cleaned up a few things.
