+++ include/ui/cli/client_request.h
@@ -17,6 +17,6 @@ void showHelpMenu();
  * Create command.
  * @param parameters list of our parameters.
  */
-void createCorpus(const std::deque<std::string>& parameters);
+void createCorpus(const std::deque<std::string> &parameters);
 
-#endif //CORPUSHARVESTER_CLIENT_REQUEST_H
+#endif // CORPUSHARVESTER_CLIENT_REQUEST_H
