+++ src/main.cpp
@@ -1,6 +1,6 @@
 #include <download/apis/ApiTwitter.h>
 #include <iostream>
-#include "utils/client_request.h"
+#include <ui/cli/client_request.h>
 
 /**
  * Get the command of the user.
@@ -11,29 +11,32 @@
  */
 int main(int argc, char **argv) {
 
-    // We have no command.
-    if (argc == 1) {
-        std::cout << "Please specify your command !" << std::endl;
-        exit(0);
-    }
+  // We have no command.
+  if (argc == 1) {
+    std::cout << "Please specify your command!" << std::endl;
+    exit(0);
+  }
 
-    // Get all the parameters pass through the command line.
-    std::deque<std::string> cmdLineArgs(argv + 1, argv + argc);
+  // Get all the parameters pass through the command line.
+  std::deque<std::string> cmdLineArgs(argv + 1, argv + argc);
 
-    // Get the first command and remove it form the deque array.
-    std::string firstCommand = cmdLineArgs.front();
-    cmdLineArgs.pop_front();
+  // Get the first command and remove it form the deque array.
+  std::string firstCommand = cmdLineArgs.front();
+  cmdLineArgs.pop_front();
 
-    // We print the help menu.
-    if (firstCommand == "help" or firstCommand == "--help" or firstCommand == "-help") {
-        showHelpMenu();
-    }
-    else if (firstCommand == "create") {
-        createCorpus(cmdLineArgs);
-    } else {
-        std::cout << "The command '" << firstCommand << "' doesn't exit. Please check the available commands with 'help' !" << std::endl;
-        exit(0);
-    }
+  // We print the help menu.
+  if (firstCommand == "help" or firstCommand == "--help" or
+      firstCommand == "-help") {
+    showHelpMenu();
+  } else if (firstCommand == "create") {
+    createCorpus(cmdLineArgs);
+  } else {
+    std::cout
+        << "The command '" << firstCommand
+        << "' doesn't exit. Please check the available commands with 'help' !"
+        << std::endl;
+    exit(0);
+  }
 
-    return 0;
+  return 0;
 }
